Create Green Urbanism Conference
The world is in the midst of a disturbance of a growing, consuming population, and environmental degradation, which produces many environmental challenges such as global climate change, excessive fossil fuel dependency and the growing demand for energy, all likely to be major challenges of the 21st century and one of the greatest problems facing humanity.

In this context, urban design is the fundamental principle of how to shape our cities. Because almost half of the energy consumed is used in cities and urban built-up areas, it is necessary to avoid mistakes in urban design at early stages in order to own more sustainable cities. The 1st international conference on Green Urbanism reflects upon practical strategies focusing on increasing sustainability beyond and within the scope of individual buildings.

The conference provides a context for a general debate about the regeneration of the city center and discusses how urbanism is affected by the paradigms of ecology.

IEREK organizes this conference to deal with cross-cut issues in architecture and urban design and address the question of how we can best cohesively integrate all the aspects of energy systems, transportation systems, waste and water management, passive and active strategies, into contemporary urban design and improved environmental performance of our cities.

The significance of this conference lies in the pressing need for the integration of sustainability principles in the urban design process of cities and the general need for a sustainable city development. It will be of particular relevance to the rapid urban growth of developing cities.

- https://www.ierek.com/events/international-conference-green-urbanism/